+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v10.1.124 (2025-05-22)
+-------------------------
+ * Update call status to in progress via sprint_ended handler+hook
+
+v10.1.123 (2025-05-21)
+-------------------------
+ * Fix IVR incoming triggers
+ * Update runner Start/Resume functions to return scenes and remove Session.sprint transient field
+
 v10.1.122 (2025-05-20)
 -------------------------
  * Move flow stat saving out of session model functions and into runner hook